Begins the road to better edit/publish
This commit admits that "cancel" isn't a function we need. We now have
"save" and "publish" as the two CTAs for the author. This implementation
is a little unintuitive as any published article is a state - published.
That is, if you publish an article, make changes to the body, and save
you will see those changes published without manually hitting the
"publish" CTA *again*.

The next step is to separate the published values versus the draft
values in the database, changing the save implementation to target
those values, and changing the publish implementation to replace the
primany values with the draft values.

E.g. "save" -> `body_draft = ${val}`
     "publish" -> `body = draft_body`